requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
groq>=0.4.0
anthropic>=0.7.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0

